commit: Remove redundant size parameter from pagination methods
This commit removes the redundant size parameter from various pagination methods in different classes. The size parameter specified the number of records per page, but it was not used in the actual implementation. By removing this parameter, we simplify the method signatures and make the code more concise.

Classes affected by this change:

    TeacherServiceImpl
    SectionController
    GradeLevelService
    TeacherService
    AttendanceWebSocketHandler
    StudentServiceImpl
    UserService
    AttendanceServiceImpl
    UserServiceImpl
    SectionService
    GuardianService