feat: updated build.gradle to use latest spring boot version and rearranged dependencies
Updated Spring Boot version from 3.3.0 to 3.3.4 to take advantage of the latest features and bug fixes. Rearranged dependencies to group similar dependencies together, such as moving developmentOnly and annotationProcessor dependencies to the end of the list. Moved runtimeOnly 'org.postgresql:postgresql' to the end of the list, as it is a runtime dependency. Added runtimeOnly 'org.springframework.boot:spring-boot-docker-compose' to the list of dependencies. This change ensures that the project is using the latest version of Spring Boot and that dependencies are organized in a logical and consistent manner.